move utils out of tests
